feat: #94 단일 시황 뉴스 조회 로직 구현

+++ back/news/src/main/java/com/ssafy/news/domain/repository/IndustryNewsRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
 
-public interface IndustryNewsRepository extends JpaRepository<IndustryNews, Integer> {
+public interface IndustryNewsRepository extends JpaRepository<IndustryNews, Long> {
     // JPA에서 이런식으로 DTO를 반환하는 건 상당히 안티패턴임
     // 그러나 현재 본문 내용이 굉장히 길기 때문에 프리뷰를 위해 모든 본문을 조회하는 건 성능 문제가 존재
     // 성능 차원에서 이렇게 반환하도록 결정함
